% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_errors.R
\name{clean_20k8k}
\alias{clean_20k8k}
\title{Clean marker data frame}
\usage{
clean_20k8k(all_data)
}
\arguments{
\item{all_data}{Marker data set as data frame}
}
\description{
Helper function; clears input marker data frame of all cells containing value "NA"
}
\examples{
clean_20k8k()
}
\keyword{data}
\keyword{input}
\keyword{marker}
\keyword{set}
