% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{cull_by_locus}
\alias{cull_by_locus}
\title{Cull by locus}
\usage{
cull_by_locus(summary)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}
}
\description{
Creates a soft subset of the marker summary frame based on marker location.
Removes redundant markers; defaults to keeping first marker at location. Returns
summary frame with column "Included" modified.
}
\examples{
cull_by_locus()
}
\keyword{data}
\keyword{loci}
\keyword{marker}
\keyword{summary}
