% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{excluded_markers}
\alias{excluded_markers}
\title{Summary-excluded Markers}
\usage{
excluded_markers(summary)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}
}
\description{
Helper function; produces a true subset of the input summary based on Included
column, values of FALSE or 0 make the subset.
}
\examples{
excluded_markers()
}
\keyword{map}
\keyword{summary}
