% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_out.R
\name{fill_pedigree}
\alias{fill_pedigree}
\title{Fill Pedigree}
\usage{
fill_pedigree(phenotype_data, all_data)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}
}
\description{
Given a phenotype and marker data, assembles a data frame with the pedigree of each input
individual represented in column 1. Each individual has their parents represented in columns
2 and 3.
}
\examples{
fill_pedigree()
}
\keyword{.dat}
\keyword{pedigree}
