% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{first_parent_SNP_data}
\alias{first_parent_SNP_data}
\title{Assemble parent marker data}
\usage{
first_parent_SNP_data(phenotype_data, all_data)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}
}
\description{
Creates a data frame subset of the curated output .dat frame, which
contains only parents of the input set, and corresponding marker data
}
\examples{
first_parent_SNP_data()
}
\keyword{data}
\keyword{marker}
\keyword{summary}
