% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_errors.R
\name{get_individuals}
\alias{get_individuals}
\title{Get (the set of) Individuals}
\usage{
get_individuals(phen_data)
}
\arguments{
\item{phen_data}{Data frame with phenotype data on all input individuals}
}
\description{
Produces a vector containing the set of all input individuals provided
the phenotype data in data frame form.
}
\examples{
get_individuals()
}
\keyword{data}
\keyword{input}
\keyword{phenotype}
\keyword{set}
