% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_out.R
\name{get_parents}
\alias{get_parents}
\title{Get Parents}
\usage{
get_parents(indiv_and_parents)
}
\arguments{
\item{indiv_and_parents}{Data frame, input individuals set and parents of each individual}
}
\description{
Given a properly-formatted data frame, with individuals in column 1, parents in columns
2 and 3, this function returns vector containing the set of all parents
}
\examples{
get_parents()
}
\keyword{.dat}
\keyword{pedigree}
