% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{included_markers}
\alias{included_markers}
\title{Summary-included Markers}
\usage{
included_markers(summary)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}
}
\description{
Helper function; produces a true subset of the input summary based on Included
column, values of TRUE or 1 make the subset.
}
\examples{
included_markers()
}
\keyword{map}
\keyword{summary}
