% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{input_SNP_summary}
\alias{input_SNP_summary}
\title{Input Set Marker Summary}
\usage{
input_SNP_summary(phenotype_data, all_data, map, start, end)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}

\item{start}{First column of data in phenotype data frame}

\item{end}{Last column of data in phenotype data frame}
}
\description{
Creates a data frame that summarizes SNP frequencies for the set of all
input individuals provided by the phenotype data.
}
\examples{
input_SNP_summary()
}
\keyword{allele}
\keyword{data}
\keyword{frequencies}
\keyword{marker}
\keyword{summary}
