% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_errors.R
\name{makeD_cleaned_input_indiv}
\alias{makeD_cleaned_input_indiv}
\title{Get "Clean" Input-Individual Set}
\usage{
makeD_cleaned_input_indiv(phenotype_data, all_data)
}
\arguments{
\item{all_data}{Marker data set as data frame}

\item{phen_data}{Data frame with phenotype data on all input individuals}
}
\description{
Using logic of function \code{makeD_get_errors()} obtains the set of
all input-individuals whose IDs exist in the master marker data set.
}
\examples{
makeD_cleaned_input_indiv()
}
\keyword{data}
\keyword{input}
\keyword{marker}
\keyword{set}
