% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_final}
\alias{map_final}
\title{Update .map}
\usage{
map_final(summary, map)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}

\item{map}{Marker map as data frame}
}
\description{
Returns data frame version of .map for use in
in FQTL, obtains image of current SNP set with group and locus data provided existing,
complete .map and input set summary. Assumes correctness of summary. Can produce default map is summary
parameter is not subset.
}
\examples{
map_format()
}
\keyword{loci}
\keyword{map}
\keyword{markers}
\keyword{summary}
