% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_format}
\alias{map_format}
\title{Format Map Helper Frame}
\usage{
map_format(map)
}
\arguments{
\item{map}{Marker map as data frame}
}
\description{
Produces a helper frame with SNP/marker location and group number paired to
each marker ID in a 3-column table
}
\examples{
map_format()
}
\keyword{map}
\keyword{markers}
\keyword{summary}
