% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{map_to_SNPs}
\alias{map_to_SNPs}
\title{Assemble marker set from map data}
\usage{
map_to_SNPs(map)
}
\arguments{
\item{map}{Data frame with map data (group number and loci)}
}
\description{
Helper function; returns vector of all SNPs/markers provided a map file as a data frame
}
\examples{
map_to_SNPs()
}
\keyword{data}
\keyword{map}
\keyword{marker}
