% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{marker_count}
\alias{marker_count}
\title{Marker Count By Group}
\usage{
marker_count(summary)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}
}
\description{
Obtains a vector of marker counts by Group number, useful to .par file.
Group number corresponds to vector index.
}
\examples{
marker_count()
}
\keyword{.par}
\keyword{constituent}
\keyword{map}
\keyword{summary}
