% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{parent_SNP_summary}
\alias{parent_SNP_summary}
\title{Parent Marker Summary}
\usage{
parent_SNP_summary(phenotype_data, all_data, map)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}
}
\description{
Creates a data frame that summarizes SNP frequencies for the set of all
parents of the input individual set provided by the phenotype data.
}
\examples{
parent_SNPs_summary()
}
\keyword{allele}
\keyword{data}
\keyword{frequencies}
\keyword{marker}
\keyword{pedigree}
\keyword{summary}
