% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{parent_SNPs_culled}
\alias{parent_SNPs_culled}
\title{Assemble parent-homozygous marker data}
\usage{
parent_SNPs_culled(phenotype_data, all_data, map)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}
}
\description{
Creates a data frame subset of the curated output .dat frame, which
contains only parents of the input set, and only homozygous markers
}
\examples{
parent_SNPs_culled()
}
\keyword{data}
\keyword{marker}
\keyword{pedigree}
