% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{subset_marker_data}
\alias{subset_marker_data}
\title{Subset Summary By Input}
\usage{
subset_marker_data(summary, map, add_remove, markers)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}

\item{map}{Marker map as data frame}

\item{add_remove}{String, "add" or "remove"}

\item{markers}{Vector of marker IDs to add or remove to input summary, Markers must be present in map set}
}
\description{
Updates input summary based on user-input string "add" or "remove"
and vector containing markers.
Best paired with functions which update dependents of the
input summary.
Requires another function to update the .dat data frame by input summary.
}
\examples{
subset_marker_data()
}
\keyword{.par}
\keyword{constituent}
\keyword{map}
\keyword{markers}
\keyword{summary}
