% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_out.R
\name{subset_out_markers}
\alias{subset_out_markers}
\title{Subset Data File}
\usage{
subset_out_markers(out, markers)
}
\arguments{
\item{out}{Data frame with phenotype data on all input individuals}

\item{markers}{Vector, set of marker IDs to subtract from .dat}
}
\description{
Helper function; obtains subset of 'out' based on a provided set of marker IDs by set subtraction
}
\examples{
subset_out_markers(out = my_dat, markers = exclude_these)
}
\keyword{.dat}
\keyword{out}
