% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{subtraction_input_summary}
\alias{subtraction_input_summary}
\title{Summary Set-Subtraction}
\usage{
subtraction_input_summary(summary, markers)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}

\item{markers}{Vector of marker IDs to exclude from returned output}
}
\description{
Soft subset of SNPs/markers included in the input summary frame by
set subtraction, summary - markers. Markers must be a vector.
}
\examples{
subtraction_input_summary(summary = your_summary_table, markers = set_to_subtract)
}
\keyword{data}
\keyword{map}
\keyword{marker}
\keyword{summary}
