% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_resolution.R
\name{summary_resolution_adjust}
\alias{summary_resolution_adjust}
\title{Adjust map resolution}
\usage{
summary_resolution_adjust(interval, map, summary)
}
\arguments{
\item{interval}{Integer sets cM interval for map resolution}

\item{map}{Data frame with map data (group number and loci)}

\item{summary}{SNP or marker summary as data frame}
}
\description{
Given an integer, adjusts map resolution and subsets input summary accordingly
}
\examples{
map_to_SNPs()
}
\keyword{data}
\keyword{map}
\keyword{marker}
