% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_errors.R
\name{to_all_indiv}
\alias{to_all_indiv}
\title{Get masterlist of indices and parents}
\usage{
to_all_indiv(all_data)
}
\arguments{
\item{all_data}{Marker data set as data frame}
}
\description{
Helper function; from input marker data frame, obtains subset containing all possible individuals,
parent 1 and parent 2 for each.
}
\examples{
to_all_indiv()
}
\keyword{data}
\keyword{input}
\keyword{marker}
\keyword{set}
