% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{to_extract_by_locus}
\alias{to_extract_by_locus}
\title{Assemble locus-culled set}
\usage{
to_extract_by_locus(map)
}
\arguments{
\item{map}{Marker map as data frame}
}
\description{
Helper function; creates a set of markers as a vector excluded by the logic of \code{cull_by_locus()}
}
\examples{
to_extract_by_locus()
}
\keyword{data}
\keyword{loci}
\keyword{locus}
\keyword{map}
\keyword{marker}
