% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_stats.R
\name{to_extract_parent_homozygous}
\alias{to_extract_parent_homozygous}
\title{Assemble set of parent-homozygous markers}
\usage{
to_extract_parent_homozygous(summary)
}
\arguments{
\item{summary}{SNP or marker summary as data frame}
}
\description{
Helper function; returns set (as vector) of all parent-homozygous SNPs/markers for
removal from curated output frame. Assumes input summary is correct.
}
\examples{
to_extract_parent_homozygous()
}
\keyword{data}
\keyword{marker}
\keyword{summary}
