% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_out.R
\name{update_out_by_summary}
\alias{update_out_by_summary}
\title{Update Data File By Summary}
\usage{
update_out_by_summary(phenotype_data, all_data, start, end, summary)
}
\arguments{
\item{phenotype_data}{Data frame with phenotype data on all input individuals}

\item{all_data}{Marker data set as data frame}

\item{start}{First column of data in phenotype data frame}

\item{end}{Last column of data in phenotype data frame}

\item{summary}{SNP or marker summary as data frame}
}
\description{
Produces a subset of all marker data based on individuals provided by the summary set.
Output may be a subset or supserset of the existing .dat data frame depending on the summary.
}
\examples{
update_out_by_summary()
}
\keyword{.dat}
\keyword{out}
